bugfix: fix settings update checker & move-notes functionality
